requirements.txt
ccxt
pandas
pandas_ta
numpy
fastapi
uvicorn
python-dotenv
pyyaml
pytest
httpx
pydantic
sqlalchemy
aiosqlite
